refactor(2TK5AD): extract connect composition root to fix cross-command import
depcruise cli-no-cross-command-imports flagged setup.ts -> connect.ts. Extract
the port-wiring (build real prompt/secret-store/verify ports + resolve env
credential + run connectTracker) into tracker-connect/run.ts. Both the connect
command and the setup offer now call runConnect with an injected logger, so
neither command imports the other. Pure refactor: behavior unchanged, the setup
offer now returns the ConnectResult directly instead of round-tripping through
process.exitCode. depcruise clean (0 violations); 19 tracker-connect + setup
tests green.
